requirements.txt
psycopg[binary]>=3.2
pgvector>=0.3
python-dotenv>=1.0
